requirements.txt
flask 
pymongo[srv]
python-dotenv
gunicorn